Update es5.md
`undefined` should be a type of Literal. Currently, there's no mention of `undefined` anywhere in the ESTree spec.

As a result, parsers like Esprima treat `undefined` as an identifier with the name "undefined," because that's the only legal way to express it given the current ESTree types at HEAD of this repo. You can see this behavior in the Esprima demo page here: http://esprima.org/demo/parse.html